Refactor list view so we can access it from ScWidget
