+++ src/openrct2-ui/scripting/CustomWindow.h
@@ -18,11 +18,14 @@
 
 namespace OpenRCT2::Ui::Windows
 {
+    class CustomListView;
+
     std::string GetWindowTitle(rct_window* w);
     void UpdateWindowTitle(rct_window* w, const std::string_view& value);
     void UpdateWidgetText(rct_window* w, rct_widgetindex widget, const std::string_view& string_view);
     rct_window* FindCustomWindowByClassification(const std::string_view& classification);
     std::optional<rct_widgetindex> FindWidgetIndexByName(rct_window* w, const std::string_view& name);
+    CustomListView* GetCustomListView(rct_window* w, rct_widgetindex widgetIndex);
 } // namespace OpenRCT2::Ui::Windows
 
 #endif
