+++ src/openrct2/scripting/ScriptEngine.cpp
@@ -1136,6 +1136,13 @@ std::string OpenRCT2::Scripting::Stringify(const DukValue& val)
     return ExpressionStringifier::StringifyExpression(val);
 }
 
+std::string OpenRCT2::Scripting::ProcessString(const DukValue& value)
+{
+    if (value.type() == DukValue::Type::STRING)
+        return language_convert_string(value.as_string());
+    return {};
+}
+
 bool OpenRCT2::Scripting::IsGameStateMutable()
 {
     // Allow single player to alter game state anywhere
